change when and where node runtime features detection happens (make ti close to what happens for browsers)
