+++ src/internal/node-launcher/node_execution_dynamic_import.js
@@ -2,10 +2,33 @@ import { resolveUrl } from "@jsenv/filesystem"
 
 import { unevalException } from "@jsenv/core/src/internal/unevalException.js"
 import { measureAsyncFnPerf } from "@jsenv/core/src/internal/perf_node.js"
+import { startObservingPerformances } from "./node_execution_performance.js"
 
-export const createNodeExecutionWithDynamicImport = ({
+export const execute = async ({
   projectDirectoryUrl,
+  fileRelativeUrl,
+  // do not log in the console
+  // because error handling becomes responsability
+  // of node code launching node process
+  // it avoids seeing error in runtime logs during testing
+  errorExposureInConsole = false,
+  collectCoverage,
+  measurePerformance,
+  collectPerformance,
 }) => {
+  let finalizeExecutionResult = (result) => result
+
+  if (collectPerformance) {
+    const getPerformance = startObservingPerformances()
+    finalizeExecutionResult = async (executionResult) => {
+      const performance = await getPerformance()
+      return {
+        ...executionResult,
+        performance,
+      }
+    }
+  }
+
   const executeFile = async (
     specifier,
     { measurePerformance, errorExposureInConsole = false } = {},
@@ -13,7 +36,6 @@ export const createNodeExecutionWithDynamicImport = ({
     // we can't dynamically import from compileServerOrigin I guess
     // we have to use the filesystem
     const fileUrl = resolveUrl(specifier, projectDirectoryUrl)
-
     const importWithDynamicImport = async () => {
       try {
         const status = "completed"
@@ -33,14 +55,20 @@ export const createNodeExecutionWithDynamicImport = ({
         }
       }
     }
-
     if (measurePerformance) {
       return measureAsyncFnPerf(importWithDynamicImport, "jsenv_file_import")
     }
     return importWithDynamicImport()
   }
 
-  return {
-    executeFile,
-  }
+  const executionResult = await executeFile(fileRelativeUrl, {
+    errorExposureInConsole,
+    measurePerformance,
+    collectCoverage,
+  })
+
+  return finalizeExecutionResult({
+    ...executionResult,
+    indirectCoverage: global.__indirectCoverage__,
+  })
 }
