--- src/nodeRuntime.js
@@ -1,5 +1,0 @@
-import { createNodeRuntime } from "./internal/runtime/createNodeRuntime/createNodeRuntime.js"
-
-export const nodeRuntime = {
-  create: createNodeRuntime,
-}
